Code Update
Inside ButtonControl.cpp, I changed one of the #define to be the pin to turn off the system.  I set it to HIGH at first, to keep the system off. Once the button is pressed for 4 seconds, it toggles the pin to low inside checkButton().

Other than that, i just added some comments and did some cleaning up on some old code that I have previously written.